Refactor selector for file count #95
